feat: Moved request types to its own file, added optional year parameter.
