+++ src/format/mod.rs
@@ -2,7 +2,7 @@ mod tana;
 pub use tana::*;
 
 mod raw;
-pub use raw::*;
+// pub use raw::*;
 
 use super::{Event, EventType};
